# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel OpenApi Blog Documentation",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "junioskyjustin@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "Blog API Server"
        }
    ],
    "paths": {
        "/categories": {
            "get": {
                "tags": [
                    "Categorie"
                ],
                "summary": "Get list of categories",
                "description": "Returns list of categories",
                "operationId": "getCategorieList",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Page not found"
                    }
                }
            },
            "post": {
                "tags": [
                    "Categorie"
                ],
                "summary": "Store new categories",
                "description": "Returns categories data",
                "operationId": "storeCategorie",
                "requestBody": {
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/categories/{id}": {
            "get": {
                "tags": [
                    "Categorie"
                ],
                "summary": "Get categories information",
                "description": "Returns categories data",
                "operationId": "getCategorieById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Categorie id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "put": {
                "tags": [
                    "Categorie"
                ],
                "summary": "Update existing categories",
                "description": "Returns updated categories data",
                "operationId": "updateCategorie",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Categorie id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true
                },
                "responses": {
                    "202": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Categorie"
                ],
                "summary": "Delete existing categories",
                "description": "Deletes a record and returns no content",
                "operationId": "deleteCategorie",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Categorie id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Blog",
            "description": "API Endpoints of Projects"
        }
    ]
}